pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meds-torch-encoders"
dynamic = ["version"]
authors = [
  {name="Matthew McDermott", email="mattmcdermott8@gmail.com"},
  {name="Nassim Oufattole", email="noufattole@gmail.com"},
  {name="Teya Bergamaschi", email="teyabergamaschi@gmail.com"},
  {name="Aleksia Kolo", email="aleksiakolo@gmail.com"},
]
description = "An efficient, flexible PyTorch dataset class for MEDS data."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "polars", "nested_ragged_tensors>=0.1.0", "numpy", "torch", "MEDS_transforms>=0.1.0", "hydra-core",
    "omegaconf", "meds-torch-data"
]

[tool.setuptools_scm]

[project.optional-dependencies]
dev = ["pre-commit<4", "ruff"]
tests = ["pytest-cov[toml]"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--doctest-modules",
  "--ignore=docs",
  "--doctest-glob=*.md",
]
doctest_optionflags = ["NORMALIZE_WHITESPACE", "ELLIPSIS"]

[project.urls]
Homepage = "https://github.com/mmcdermott/meds-torch-encoders"
Issues = "https://github.com/mmcdermott/meds-torch-encoders/issues"

[tool.ruff]
target-version = "py312"
line-length = 110

# Enable linters
[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "RUF", # Ruff-specific rules
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "N",   # pep8-naming
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__ files
"test_*" = ["DTZ001"]  # Ignore datetimez in test files

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["*MEDS*", "*NRT*"]
